requirements.txt
pyyaml
pandas
numpy
akshare
ccxt
requests
